requirements.txt
# 框架依赖
paddlepaddle>=3.0.0,<4.0.0
paddledet>=2.6.0,<3.0.0  # PaddleDetection新版本包名
albumentations>=1.3.0,<2.0.0

# 数据处理
opencv-python>=4.6.0,<5.0.0
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
scipy>=1.7.0,<2.0.0
Pillow>=9.0.0,<10.0.0
imgaug>=0.4.0,<1.0.0

# 配置和日志
PyYAML>=6.0.0,<7.0.0
visualdl>=2.5.0,<3.0.0
tqdm>=4.62.0,<5.0.0
matplotlib>=3.5.0,<4.0.0

# 数据标注工具
labelme>=5.1.1,<6.0.0

# 测试工具
pytest>=7.0.0,<8.0.0
pytest-cov>=3.0.0,<4.0.0
pytest-benchmark>=4.0.0,<5.0.0

# 代码质量工具
flake8>=4.0.0,<5.0.0
black>=22.0.0,<23.0.0
isort>=5.10.0,<6.0.0

# 性能优化
psutil>=5.9.0,<6.0.0  # 系统资源监控